requirements.txt
torch
torchvision
gnutools-python
